+++ qs-stdw-wallet/src/lattice/poly_from_seed.h
@@ -5,10 +5,9 @@
 #include "../../config/params.h"
 
 /*
- * Deterministically expand a seed into a polynomial
- * in R_q with coefficients derived via SHAKE256.
+ * Deterministically expand a seed into a polynomial in R_q.
+ * Coefficients are in [0, q-1] where q = RACCOON_Q (49-bit composite).
  */
-void poly_from_seed(int32_t a[QS_N],
-                    const uint8_t seed[SEED_BYTES]);
+void poly_from_seed(int64_t a[QS_N], const uint8_t seed[SEED_BYTES]);
 
 #endif
